Set NODE_ENV and only add purgecss in production
